Issue 1460 Subspec dependency checking should no longer produce false-positives
Previously, when defining dependencies in a pod spec, an error would be thrown if
the parent spec's name includes the dependency name (because it thinks a cyclic
dependency is created). For example, a pod spec named 'SpecTest' would incorrectly
fail if one of it's sub specs had a dependency on a pod named 'Spec'.